+++ internal/sema/symbols.go
@@ -47,6 +47,7 @@ const (
 
 type fieldSymbol struct {
 	static          bool
+	declaration     *ast.FieldDecl
 	typeInfo        Type
 	visibility      ast.Visibility
 	goName          string